requirements.txt
Django>=1.3.0
mimeparse>=0.1.3
python-dateutil>=2.1
lxml
defusedxml
PyYAML
python-digest
biplist